feat: add Menu button and frontend modularization
