--- client/containers/SearchContainer.jsx
@@ -1,1 +1,0 @@
-// NOT UTILIZED
