+++ client/src/components/Header.jsx
@@ -1,9 +1,15 @@
 import React from 'react';
+import Menu from './Menu';
 
-function Header() {
+function Header({ isOpen, setIsOpen, handleMenuBtn }) {
   return (
-    <div id="browserLogo">
-      <img src="/Shirt Logo Draft.png" alt="Logo" />
+    <div id='browserLogo'>
+      <img src='/Shirt Logo Draft.png' alt='Logo' />
+      <Menu
+        handleMenuBtn={handleMenuBtn}
+        setIsOpen={setIsOpen}
+        isOpen={isOpen}
+      />
       {/* Optionally, include more images or elements here */}
       {/* <img src="/l-intro-1630426166.jpg" alt="pic1" />
       <img src="/istockphoto-1322887164-612x612.jpg" alt="pic2" />
